requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2